beautify code
