+++ base/file_utils/BUILD
@@ -2,6 +2,7 @@ py_library(
     name = "file_utils",
     srcs = [
         "file_utils.py",
+        "lock_file.py",
     ],
     visibility = ["//visibility:public"],
 )
